beat-aligned effect freqency mods were out of sync, due to non-fractional sample counting (samplesPerTick is commonly a non-integer)

+++ app/src/main/jni/ticker.c
@@ -12,7 +12,7 @@ jlong Java_com_odang_beatbot_manager_MidiManager_getCurrTick(JNIEnv *env,
 void Java_com_odang_beatbot_manager_MidiManager_setNativeMSPT(JNIEnv *env,
                                                               jclass clazz, jlong MSPT) {
     // MSPT = microseconds-per-tick
-    samplesPerTick = (MSPT * SAMPLE_RATE) / 1000000;
+    samplesPerTick = ((float) MSPT * SAMPLE_RATE) / (float) 1000000;
 }
 
 void Java_com_odang_beatbot_manager_MidiManager_setLoopTicksNative(JNIEnv *env,
